requirements.txt
google-generativeai
PyMuPDF
pymupdf4llm
tqdm
python-dotenv